// packages/impromat/src/store/schema.gen.ts
/* tslint:disable */
/**
 * This file was automatically generated by json-schema-to-typescript.
 * DO NOT MODIFY IT BY HAND. Instead, modify the source JSONSchema file,
 * and run json-schema-to-typescript to regenerate this file.
 */

/**
 * An improvisational theatre Workshop that a person can hold for their group.
 */
export interface Workshop {
  /**
   * For syncing purposes this flag indicates deleted workshops.
   */
  deleted?: boolean | null;
  description: string;
  id: string;
  name: string;
  sections: Section[];
  updatedAt: unknown;
}
/**
 * A section groups several elements of a workshop. A workshop must always have at least one section.
 */
export interface Section {
  color?: string | null;
  elements: Element[];
  id: string;
  isCollapsed?: boolean | null;
  /**
   * If false the section will not be shown. If there are no sections yet there is an invisiable ghost section to hold all elements.
   */
  isVisible?: boolean | null;
  name: string;
  note?: string | null;
}
/**
 * General element like a exercise or game.
 */
export interface Element {
  basedOn?: Element | null;
  id: string;
  languageCode?: string | null;
  licenseName?: string | null;
  licenseUrl?: string | null;
  markdown: string;
  name: string;
  /**
   * User defined note
   */
  note: string;
  sourceBaseUrl?: string | null;
  sourceName?: string | null;
  sourceUrl?: string | null;
  tags: string[];
}